Upgrade openssl to 1.0.2
Summary:
Upgrade libcrypto to the 1.0.2
version 4ac0329582829f5378d8078c8d314ad37db87736.

This is a slight increase in binary
size over the previous version, since
I have tried to keep source code
modifications to a minimum.

We're also changing the way we're managing
openssl. We're going to start checking in
openssl as prebuilt binaries instead of
checking in the source and we'll still post
patches so that you can build your own
openssl with the magic parameter incantation.

This change will allow us to more quickly
respond to openssl upgrades so that we can
keep up-to-date.

Test Plan:
Ran tests on armeabi devices, armeabi-v7a devices
as well as an x86 emulator, they all work as
expected.